refactor: move tsky class into own subfile and export all files from index
